+++ R/spades-project-package.R
@@ -5,8 +5,11 @@
 #' \if{html}{\figure{SpaDES.png}{options: width=100 alt="SpaDES logo" style="float: right;"}}
 #' \if{latex}{\figure{SpaDES.png}{options: width=0.5in}}
 #'
-#' Quickly setup 'SpaDES' project directories and add modules using templates. 
+#' Quickly setup 'SpaDES' project directories and add modules using templates.
 #'
 #' @import methods
 #' @rdname SpaDES.project-package
 "_PACKAGE"
+
+
+linkOrCopy <- getFromNamespace("linkOrCopy", ns = "Require")